libxfce4ui: Change vendor info
debian/rules passes value of XFVENDOR to the package's configure script
as the value of the "--with-vendor-info" option.  This sets the string
in the "distributed by" message in the xfce4-about program.

debian/vendorinfo also needs to be changed because debian/rules installs
this file to /usr/share/xfce4/vendorinfo if the value of XFVENDOR is not
"Xubuntu".

Fixes issue #15432.