Add helpers for linux-lts-xenial and linux-meta-lts-xenial
Mostly based off flidas linux and linux-meta helpers.

ZFS is not built by default for linux-lts-xenial upstream, but I removed the
code and supporting files for it anyway.

The package helper modifies files in the debian.master directory, even though
I believe that directory is not used. This is done to remove any lingering
references to non-free software in the package, and to make comparing the
debian.master and debian.xenial directories less confusing if anyone wants to
do that.