Remove backport for ansible package - renaming makes it unworkable
Ansible has not published a package for focal in the ppa due to
multiple build issues[1].

What exists for focal is ansible-base: the ansible project was
restructured and split into ansible-base (later to be ansible-core)
and a collection of modules and plugins which is now ansible[2].
In principle, for future releases ansible package would contain the
collections and depend on ansible-base/ansible-core, which would replace
the current ansible package. However, that work has not been done in the
ppa, which neither provides the collections.

[1] https://github.com/ansible/ansible/issues/69203
[2] https://www.ansible.com/blog/ansible-3.0.0-qa