Remove partial rebranding of ubuntustudio packages
Ubuntustudio packages are purged from upstream for branding
reasons. It includes a great amount of packages, out of which
there exist multiple interrelated dependencies. If somebody
is ever interested in working on the branding, then all of
them should be rebranded and ported together. Otherwise,
it will not be possible to install some of them, which is
currently the case with trisquelstudio-menu.